Support implicit IP Address of client
Often, if this is being used on the client we are wanting information about the current user. On the client we do not have access (within the browser) to the users IP address. The 3 default providers support passing no IP address, this will then return the location details for the current user.

By omitting the IP address parameter (or `True` if you wish to imply that true means use current users IP address) this will now return information about the current user.